% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files.R
\name{list_files}
\alias{list_files}
\title{List DATIM files in a directory}
\usage{
list_files(path = "data/", full.names = TRUE)
}
\arguments{
\item{path}{A character string specifying the path to the directory. Defaults to "data/".}

\item{full.names}{A logical value indicating whether to return the full file paths. Defaults to `TRUE`.}
}
\value{
A character vector of file paths.
}
\description{
This function lists all DATIM files in the specified "data/" directory.
It provides the full file paths, making it easier to reference files for further processing.
}
\examples{
# List all files in the "data/" directory
list_files()

# Specify a custom directory
list_files(path = "custom_directory/")

}
